move arcgis-rest-service-admin into arcgis-rest-features
