+++ packages/arcgis-rest-features/src/getLayer.ts
@@ -1,8 +1,8 @@
 /* Copyright (c) 2018 Environmental Systems Research Institute, Inc.
  * Apache-2.0 */
 
-import { request, cleanUrl, ILayerDefinition } from "@esri/arcgis-rest-request";
-import { IGetLayerOptions } from "./helpers.js";
+import { request, cleanUrl } from "@esri/arcgis-rest-request";
+import { IGetLayerOptions, ILayerDefinition } from "./helpers.js";
 
 /**
  * ```js
