+++ packages/arcgis-rest-features/src/getServiceAdminInfo.ts
