+++ packages/arcgis-rest-features/src/getViewSources.ts
