+++ packages/arcgis-rest-features/test/addToServiceDefinition.test.ts
@@ -6,11 +6,10 @@ import { TOMORROW } from "../../../scripts/test-helpers.js";
 import {
   encodeParam,
   ErrorTypes,
-  UserSession,
-  ITable
+  UserSession
 } from "@esri/arcgis-rest-request";
-import { ILayer } from "../src/helpers.js";
-import { addToServiceDefinition } from "../src/addTo.js";
+import { ILayer, ITable } from "../src/helpers.js";
+import { addToServiceDefinition } from "../src/addToServiceDefinition.js";
 import {
   AddToFeatureServiceSuccessResponseFredAndGinger,
   AddToFeatureServiceSuccessResponseFayardAndHarold,
