+++ packages/arcgis-rest-features/test/mocks/layerDefinition.ts
@@ -1,7 +1,8 @@
 /* Copyright (c) 2018 Environmental Systems Research Institute, Inc.
  * Apache-2.0 */
 
-import { ILayerDefinition, IExtent } from "@esri/arcgis-rest-request";
+import { IExtent } from "@esri/arcgis-rest-request";
+import { ILayerDefinition } from "../../src/helpers.js";
 
 const defaultExtent: IExtent = {
   xmin: -180,
