--- packages/arcgis-rest-service-admin/package.json
@@ -1,69 +1,0 @@
-{
-  "name": "@esri/arcgis-rest-service-admin",
-  "version": "3.3.0",
-  "description": "Service administration helpers for @esri/arcgis-rest-js",
-  "license": "Apache-2.0",
-  "keywords": [
-    "ES6",
-    "arcgis",
-    "esri",
-    "fetch",
-    "promise",
-    "typescript"
-  ],
-  "type": "module",
-  "main": "dist/cjs/index.js",
-  "module": "dist/esm/index.js",
-  "unpkg": "dist/bundled/service-admin.umd.min.js",
-  "exports": {
-    ".": {
-      "import": "./dist/esm/index.js",
-      "require": "./dist/cjs/index.js"
-    }
-  },
-  "types": "dist/esm/index.d.ts",
-  "sideEffects": false,
-  "files": [
-    "dist/**"
-  ],
-  "scripts": {
-    "build": "npm-run-all --parallel build:*",
-    "postbuild": "node ../../scripts/create-dist-package-jsons.js",
-    "build:bundled": "rollup -c ../../rollup.js",
-    "build:cjs": "tsc --outDir ./dist/cjs -m commonjs",
-    "build:esm": "tsc --outDir ./dist/esm --declaration",
-    "dev": "npm-run-all --parallel dev:*",
-    "dev:bundled": "rollup -w -c ../../rollup.js",
-    "dev:cjs": "tsc -w --outDir ./dist/cjs -m commonjs",
-    "dev:esm": "tsc -w --outDir ./dist/esm --declaration"
-  },
-  "engines": {
-    "node": ">=12.20.0"
-  },
-  "dependencies": {
-    "tslib": "^2.3.0"
-  },
-  "peerDependencies": {
-    "@esri/arcgis-rest-portal": "^3.0.0",
-    "@esri/arcgis-rest-request": "^3.0.0"
-  },
-  "devDependencies": {
-    "@esri/arcgis-rest-portal": "^3.3.0",
-    "@esri/arcgis-rest-request": "^3.3.0"
-  },
-  "contributors": [
-    "Mike Tschudi <mtschudi@esri.com>"
-  ],
-  "homepage": "https://github.com/Esri/arcgis-rest-js#readme",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/Esri/arcgis-rest-js.git",
-    "directory": "packages/arcgis-rest-service-admin"
-  },
-  "bugs": {
-    "url": "https://github.com/Esri/arcgis-rest-js/issues"
-  },
-  "publishConfig": {
-    "access": "public"
-  }
-}
