+++ typedoc.json
@@ -9,7 +9,6 @@
     "./packages/arcgis-rest-geocoding/",
     "./packages/arcgis-rest-portal/",
     "./packages/arcgis-rest-request/",
-    "./packages/arcgis-rest-routing/",
-    "./packages/arcgis-rest-service-admin/"
+    "./packages/arcgis-rest-routing/"
   ]
 }
